+++ lib/ui/pages/recipe/recipe-details-page.dart
@@ -56,8 +56,7 @@ class _RecipesDetailsPageState extends State<RecipesDetailsPage> {
         var ingredient = recipe.ingredients[i];
         var ingredientCpy = ingredientsTmp[i];
         ingredientCpy.amount = ((ingredient.amount / recipe.numberOfPersons) *
-                numberOfPersonsTmp)
-            .toInt();
+                numberOfPersonsTmp);
       }
     });
   }
@@ -70,8 +69,7 @@ class _RecipesDetailsPageState extends State<RecipesDetailsPage> {
         var ingredient = recipe.ingredients[i];
         var ingredientCpy = ingredientsTmp[i];
         ingredientCpy.amount = ((ingredient.amount / recipe.numberOfPersons) *
-                numberOfPersonsTmp)
-            .toInt();
+                numberOfPersonsTmp);
       }
     });
   }
